+++ scripts/dp-bin-dir2edt.cmd
@@ -6,8 +6,8 @@ rem %2 - path to folder to save 1C data processors & reports in 1C:EDT format
 rem %3 - path to 1C configuration (binary (*.cf), 1C:Designer XML format or 1C:EDT format)
 rem      or folder contains 1C infobase used for convertion
 
-if not defined V8_VERSION set V8_VERSION=8.3.20.2290
-if not defined V8_TEMP set V8_TEMP=%TEMP%\1c
+IF not defined V8_VERSION set V8_VERSION=8.3.20.2290
+IF not defined V8_TEMP set V8_TEMP=%TEMP%\1c
 
 set V8_TOOL="C:\Program Files\1cv8\%V8_VERSION%\bin\1cv8.exe"
 FOR /F "usebackq tokens=1 delims=" %%i IN (`where ring`) DO (
@@ -20,21 +20,21 @@ set WS_PATH=%V8_TEMP%\edt_ws
 set CLEAN_AFTER_EXPORT=0
 
 set DP_BIN_PATH=%1
-if defined DP_BIN_PATH set DP_BIN_PATH=%DP_BIN_PATH:"=%
+IF defined DP_BIN_PATH set DP_BIN_PATH=%DP_BIN_PATH:"=%
 set DP_SRC_PATH=%2
-if defined DP_SRC_PATH set DP_SRC_PATH=%DP_SRC_PATH:"=%
+IF defined DP_SRC_PATH set DP_SRC_PATH=%DP_SRC_PATH:"=%
 set BASE_CONFIG=%3
-if defined BASE_CONFIG set BASE_CONFIG=%BASE_CONFIG:"=%
+IF defined BASE_CONFIG set BASE_CONFIG=%BASE_CONFIG:"=%
 
-if not defined DP_BIN_PATH (
+IF not defined DP_BIN_PATH (
     echo Missed parameter 1 "path to folder contains data processors & reports in binary format (*.epf, *.erf)"
     exit /b 1
 )
-if not defined DP_SRC_PATH (
+IF not defined DP_SRC_PATH (
     echo Missed parameter 2 "path to folder to save 1C data processors & reports in 1C:EDT format"
     exit /b 1
 )
-if not exist "%BASE_CONFIG%" (
+IF not exist "%BASE_CONFIG%" (
     echo Path "%BASE_CONFIG%" doesn't exist ^(parameter 3^), empty infobase will be used.
     set BASE_CONFIG=
 )
@@ -45,20 +45,20 @@ IF "%BASE_CONFIG%" equ "" (
     set CLEAN_AFTER_EXPORT=1
     set BASE_CONFIG_DESCRIPTION=empty configuration
     %V8_TOOL% CREATEINFOBASE File=%IB_PATH%; /DisableStartupDialogs
-) else (
+) ELSE (
     set BASE_CONFIG_DESCRIPTION=configuration from "%BASE_CONFIG%"
     IF exist "%BASE_CONFIG%\DT-INF\" (
         set CLEAN_AFTER_EXPORT=1
         call %~dp0edt2ib.cmd "%BASE_CONFIG%" "%IB_PATH%"
-    ) else (
+    ) ELSE (
         IF exist "%BASE_CONFIG%\Configuration.xml" (
             set CLEAN_AFTER_EXPORT=1
             call %~dp0xml2ib.cmd "%BASE_CONFIG%" "%IB_PATH%"
-        ) else (
+        ) ELSE (
             IF exist "%BASE_CONFIG%\1cv8.1cd" (
                 set BASE_CONFIG_DESCRIPTION=existed configuration
                 set IB_PATH=%BASE_CONFIG%
-            ) else (
+            ) ELSE (
                 set CLEAN_AFTER_EXPORT=1
                 call %~dp0cf2ib.cmd "%BASE_CONFIG%" "%IB_PATH%"
             )
@@ -67,18 +67,10 @@ IF "%BASE_CONFIG%" equ "" (
 )
 
 echo Clear temporary files...
-IF "%CLEAN_AFTER_EXPORT%" equ "1" (
-    rd /S /Q "%IB_PATH%"
-)
-if exist "%XML_PATH%" (
-    rd /S /Q "%XML_PATH%"
-)
-if exist "%WS_PATH%" (
-    rd /S /Q "%WS_PATH%"
-)
-if exist "%DP_SRC_PATH%" (
-    rd /S /Q "%DP_SRC_PATH%"
-)
+IF "%CLEAN_AFTER_EXPORT%" equ "1" IF exist "%IB_PATH%" rd /S /Q "%IB_PATH%"
+IF exist "%XML_PATH%" rd /S /Q "%XML_PATH%"
+IF exist "%WS_PATH%" rd /S /Q "%WS_PATH%"
+IF exist "%DP_SRC_PATH%" rd /S /Q "%DP_SRC_PATH%"
 md "%V8_TEMP%"
 md "%XML_PATH%"
 md "%WS_PATH%"
@@ -103,12 +95,6 @@ echo Export dataprocessors ^& reports from 1C:Designer XML format "%XML_PATH%" t
 call %RING_TOOL% edt workspace import --project "%DP_SRC_PATH%" --configuration-files "%XML_PATH%" --workspace-location "%WS_PATH%" --version "%V8_VERSION%"
 
 echo Clear temporary files...
-IF "%CLEAN_AFTER_EXPORT%" equ "1" (
-    rd /S /Q "%IB_PATH%"
-)
-if exist "%XML_PATH%" (
-    rd /S /Q "%XML_PATH%"
-)
-if exist "%WS_PATH%" (
-    rd /S /Q "%WS_PATH%"
-)
+IF "%CLEAN_AFTER_EXPORT%" equ "1" IF exist "%IB_PATH%" rd /S /Q "%IB_PATH%"
+IF exist "%XML_PATH%" rd /S /Q "%XML_PATH%"
+IF exist "%WS_PATH%" rd /S /Q "%WS_PATH%"
